Make stamina consumption classes public so others can use.
