+++ src/main/java/emu/grasscutter/game/managers/StaminaManager/StaminaManager.java
@@ -1,4 +1,4 @@
-package emu.grasscutter.game.managers;
+package emu.grasscutter.game.managers.StaminaManager;
 
 import emu.grasscutter.Grasscutter;
 import emu.grasscutter.game.entity.EntityAvatar;
@@ -35,48 +35,8 @@ public class StaminaManager {
     private int staminaRecoverDelay = 0;
     private boolean isInSkillMove = false;
 
-    private enum ConsumptionType {
-        None(0),
 
-        // consume
-        CLIMB_START(-500),
-        CLIMBING(-150),
-        CLIMB_JUMP(-2500),
-        SPRINT(-1800),
-        DASH(-360),
-        FLY(-60),
-        SWIM_DASH_START(-200),
-        SWIM_DASH(-200),
-        SWIMMING(-80),
-        FIGHT(0),
 
-        // restore
-        STANDBY(500),
-        RUN(500),
-        WALK(500),
-        STANDBY_MOVE(500),
-        POWERED_FLY(500);
-
-        public final int amount;
-
-        ConsumptionType(int amount) {
-            this.amount = amount;
-        }
-    }
-
-    private class Consumption {
-        public ConsumptionType consumptionType;
-        public int amount;
-
-        public Consumption(ConsumptionType ct, int a) {
-            consumptionType = ct;
-            amount = a;
-        }
-
-        public Consumption(ConsumptionType ct) {
-            this(ct, ct.amount);
-        }
-    }
 
     public boolean getIsInSkillMove() {
         return isInSkillMove;
